feat: move configs in separate folder
