+++ src/main/java/org/example/configs/SchedulerConfig.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.example.configs;
 
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.context.annotation.Configuration;
